remove listeners & add step to push to s3
